LU-1269 ldlm: speed up ASTs sending
ldlm_run_bl_ast_work() sends ASTs in set of PARALLEL_AST_LIMIT
requests and waits for whole set to complete and then sends another
set of requests and waits again. If there is a least one request per
set which timeouts, we have timeout serialization.

This patch changes ldlm_run_bl_ast_work() so that having sent one
set it then waits for any of its requests to complete and refills
the running set with requests which are yet to be sent. For a case
where number of timeout-ing requests is smaller than
PARALLEL_AST_LIMIT it is supposed to eliminate possibility of timeout
serailization.

This patch uses posibility to specify wait condition for
ptlrpc_set_wait().

Xyratex-bug-id: MRP-478

Reviewed-by: Iurii Golovach <iurii_golovach@xyratex.com>

Signed-off-by: Vladimir Saveliev <vladimir.saveliev@oracle.com>
Change-Id: Ic05fc8f04ed90c147ca388d1e7826de4cf497799
Signed-off-by: Iurii Golovach <iurii_golovach@xyratex.com>