LU-249 mds: Increase the maximum number of threads on MDS
Currently the largest number of the threads that can exist on
the MDS is 512. For very large system is those threads can
easily be exhausted and end up back logging the system. This
patch allows the user to define the upper limit of how many
threads are created on the MDS. If not selected the default
behavior of 512 is kept.

Change-Id: Ie48eb00f48ddcd98cc518178ba5e1da9ea1e40eb
Signed-off-by: James Simmons <uja.ornl@gmail.com>