OSF-69 dne: add DNE architecture
The patch is too big to be reviewed now, and it needs to be
split. Land it just for some tests.

1. Add DNE architecture based on LOD/OSP. OSP will be used
to communicate between MDT and MDT. On the remote MDT, OUT
(object update target) will be used to handle the object
updates between MDT, which talks to OSD directly.

3. Some mgs llog fixes for DNE.
Change-Id: If6be9dcaedb01e8fcd93a0d14389582a06673759
Signed-off-by: Wang Di <di.wang@whamcloud.com>