LU-1257 mgs: treat -EADDRINUSE as non-fatal error
Sometimes, server registration could be successful on MGS,
but the registration RPC returned error because of the MGS
revokeing config locks timeout or networking problems, that
will cause the server can never be registered properly
without erasing all server configs.

This patch treats -EADDRINUSE as an non-fatal error for the
registration, and prints error messages on the console of
both MGS and target server for user to make decision.

Signed-off-by: Niu Yawei <niu@whamcloud.com>
Change-Id: If9e4e56912bbac6eee333c958f191f0d18015f6f