LU-1030 osc: support to change max_pages_per_rpc
Changing max_pages_per_rpc is not easy as before any more after
osc_extent is used because each extent relies on that value to make
the boundary correctly.

In this patch, when changing max_pages_per_rpc, it's allowed for
`old' extents and `new' extents to co-exist for a short while, and
the greater one will be used to compose RPCs.

Signed-off-by: Jinshan Xiong <jinshan.xiong@whamcloud.com>
Change-Id: Ic157e9c9c117fb4ec029c2edfbc5021b5e993f47