LU-1199 build: lustre spec file overhaul
Currently a disconnect exist between the rpm build and autoconf
system. For example the autoconf system you can over ride the
defaults, i.e prefix, where the software stack will be installed
on an given system image. When building a rpm it is expected to
pass those configure options by the define construct for rpmbuild.
Currently Lustre does not respect that. This patch address this.
This patch always you to define the location where the kernel
modules and mount.lustre can be.

Change-Id: I49e8b3b9fcd8edc52f93075032da96e248b1056f
Signed-off-by: James Simmons <uja.ornl@gmail.com>