LU-445 lnet: Send timestamps with LNet counters
It's backport of patch on LU-445.

We use a millisecond timer on a per-session basis. Each node then
sends the milliseconds since session start over with the LNet
counters. With 32bits and millisecond resolution, this should
allow ~49 days of session age before the timers would wrap.

This will change the wire attribute and not backward compatible.

Signed-off-by: Liang Zhen <liang@whamcloud.com>
Change-Id: Ied0df42634e1673b4d9e53846b60558f0042bbfd