LU-1391 lov: Hold LOV ref count until target OSCs are released
When LOV obd_device on a client node is registered with target OSCs
as observer, LOV obd_refcount isn't incremented at the time.
Thus, there is a possibility that obd_observer could be freed while
still be referred, like in obd_nitify_observer().

This patch increments/decrements LOV obd_refcount with
obd_register/unregister_observer_with_osc() when LOV is
registered/unregistered as observer. And the patch adds obd_putref()
and obd_getref() into client_disconnect_export() not to call the
function again in itself. (it's end up in a dead-lock of
cli->cl_sem without it)

Signed-off-by: Hiroya Nozaki <nozaki.hiroya@jp.fujitsu.com>
Change-Id: Ib4d34c10d44b430e79a65ebdd7e4402e5f2c8c31