LU-1095: debug: Clarify a IR message about support.
Durning testing of Imperative Recovery we turned off IR but in the
logs we saw a IR:ON message which is confusing since we set it to
disable on the MGS. The IR:ON message is only to make aware that
IR functionality exist. A better message would be IR: [Un]Supported.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: I5990b26b326ce912faf6c5d556cb92334748322c