LU-1199 build : Pass in the configure arguments to rpmbuild.
Currently the rpm spec file for lustre and ldiskfs parses
the configure arguments inside the rpm spec file instead
of the more traditional way of passing arguments with
rpmbuild. This patch is the first step to doing this by
parsing the configure options in the Makefile and then
calling rpmbuild in the Makefile with rpmbuild options
which the spec file can act upon. This patch as a first
step converts the kernel source and object options of
configure into rpmbuild options and uses them in the
spec file.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: I49e8b3b9fcd8edc52f93075032da96e248b1056f