LU-1387 lvfs: print message on allocation failure
Back in Bugzilla b=12211 the introduction of the "conditionally
fail allocations" feature, the printing of error messages on the
console for memory allocation failures was accidentally disabled.
This was a result of subtle boolean/precedence changes, and not
intentional.

Reinstate the printing of console errors on allocation failures
with an improved format that prints the actual location instead
of obd_alloc_fail().

Fix a bug in null_alloc_rs() that was hit while testing this code.

Test-Parameters: fortestonly
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I4ea2b93d2b11869940d687babd2eeafd08ca500c