LU-1387 lvfs: print message on allocation failure
Back in commit ca6dd9a48e7b3ebf5df8a78c8ed65f7676cebf4d (b=12211),
the introduction of the "conditionally fail allocations" feature
cause the printing of error messages on the console for memory
allocation failures to be accidentally disabled.  This was a result
of subtle boolean/precedence changes, and not intentional.

Reinstate the printing of console errors on allocation failures
with an improved format that prints the actual location instead
of obd_alloc_fail().

During testing of the conditional memory allocation failure code,
there were several LASSERT() failures in null_alloc_rs().  While
the core bug was not yet fixed, improve the LASSERT() verbosity for
debugging any problem hit there in the future.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I4ea2b93d2b11869940d687babd2eeafd08ca500c