LU-1497 client: SLES11 SP0 client build is broken.
With the addition of client improvements new linux kernel
apis are exposed to the llite layer. Those apis can be
missing or have a different format in earlier kernels.
This is the case for SLES11 SP0 clients whose
vma handling is different than later kernel version.
This patch fixes those issues.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: Id429a7fc5a16dafecba4ac074b7f815811e1c611