LU-1440 tests: restore remount-ro flag to local.sh
The remount-ro flag was removed from the --mkfsoptions supplied for
the MDT, and this caused a warning from mkfs.lustre.  However, this
isn't necessarily a valid fix for this issue.  It would be better if
mkfs_lustre.c always added errors=remount-ro to the specified options
if no other errors={panic,continue} flag was given.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I89c21258c74f5abcb71c4d0446f0e9cf918cfc70