LU-1432 ptlrpc: LBUG in lprocfs_free_client_stats()
When target_handle_connect() does reconnection with an existing export
object obtained from lustre_hash_lookup(), there is a case that the
refcount of the export object reaches zero while still be used there.
This ends up two phenomena the one is causing OS-panic, the other is
the LBUG in lprocfs_free_client_stats(), because the nid_stat can be
attached to the export object that was already finalized by
obd_cleanup().

This patch adjusts the refcount of export in target_handle_connect()
properly in order not to cause the above phenomena.

Signed-off-by: Hiroya Nozaki <nozaki.hiroya@jp.fujitsu.com>
Change-Id: Ibc3caf7e1426a3a99f059df0faabd7bb7314bbe8