LU-1538 tests: cleanup test output, acc-small
There are a number of issues observed when running test scripts
on a single-node system.  Several minor cleanups to the test
scripts are also possible (coding style, more clear output, etc):

- sanity.sh test_51ba() should clean up subdirectories even if
  test_51b() failed to create all of the files
- sanity.sh test_61ic() shouldn't print all files to stdout

Implement acceptance-small.sh via auster, so that we don't have to
maintain two separate test wrappers, and it would be possible for
users running acc-sm to upload results to Maloo.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I8aae9c1af508d8173135f0938e0a26d9731253df