LU-519 Asserting on a possible situation in o2iblnd_cb.c
A customer has run into an assert in kiblnd_tx_complete():
LASSERT(tx->tx_sending > 0).  It looks like their IB driver
"may" issue a complete even when returning an -ENOMEM error.

The above assert is not expecting this to happen.  This fix
is meant to address this unique situation so we don't assert
on this completion callback, but just close the connection
due to the error in hopes it can be recovered and re-established.

Note: getting -ENOMEM from the driver may be an error in
the driver itself.  This fix is not addressing that.

Change-Id: Ie7fd3980db725ce10886667206b7459ead03c98e
Signed-off-by: Doug Oucharek <doug@whamcloud.com>