LU-1432 ptlrpc: LBUG in lprocfs_free_client_stats()
When target_handle_connect() does reconnection with an existing export
object obtained from lustre_hash_lookup(), there is a case that the
refcount of the export object reaches zero while still be used there.
This ends up in two phenomena the one is causing OS-panic when the
expor object has been already freed, the other is the LBUG in
lprocfs_free_client_stats() when the export object has been already
listed in obd_zombie_exports.

This patch adjusts the refcount of export in target_handle_connect()
properly in order not to cause the above phenomena and serialize
class_cleanup() and target_handle_connect() in order to remove the
race-condition of obd_nid_hash

Signed-off-by: Hiroya Nozaki <nozaki.hiroya@jp.fujitsu.com>
Signed-off-by: Lai Siyao <laisiyao@whamcloud.com>
Change-Id: Ibc3caf7e1426a3a99f059df0faabd7bb7314bbe8