LU-398 ptlrpc: Add NRS and libcfs_heap
The Network Request Scheduler (NRS) allows a user to control the way
in which RPCs are dispatched from PTLRPC services. This can be used to
various effects, most importantly improved fairness across the
cluster, better I/O performance under certain workloads, and allowing
for QoS semantics to be used at the filesystem level. Central to the
operation of NRS are NRS policies, each of which implements a
different 'strategy', or 'personality', that affects the dispatching
of RPCs in a different way, aiming to achieve a particular goal.

This patch adds a core NRS framework, some initial policies, and a
binary heap implementation that policies make use of in order to
maintain prioritized queues of large numbers of RPCs at server nodes.
The policies currently available are:
- FIFO (First-In, First-Out),
this is a logical wrapper around previous PTLRPC
functionality; it is used as the default policy for all RPC
types on all services, and also as a fallback policy in case
one of the other policies fails to handle a given RPC.
- CRR-E (Client Round Robin over Exports),
this policy handles RPCs in a round robin fashion across
available exports seen at the server. It's aim is to provide
for improved fairness through better load balancing and
utilization of cluster resources.
- CRR-N (Client Round Robin over NIDs),
this policy handles RPCs in a round robin fashion across
available NIDs seen at the server; it's aim is the same as for
CRR-E. For clients with a single mount of the filesystem,
CRR-E and CRR-N behave identically.
- ORR (Object-based Round Robin),
this policy operates only on brw RPCs; the policy dispatches
the RPCs in groupings based on the object ID of the backend-fs
target object for the RPC. The aim of this policy is mainly to
provide for increased read throughput by minimizing the
overall amount of time spent performing disk seeks; this is
achieved by dispatching read RPCs in accordance to the order
in which the objects have been layed down on disk. Increased
read throughput using ORR may also be achieved (perhaps at the
expense of write throughput), if OST_WRITE RPCs are aligned
by logical offset before they are dispatched (more information
on this can be found below).
-TRR (Target-based Round Robin),
this policy works in a similar fashion to ORR; it also
operates only on brw RPCs, but instead of object IDs, it
groups brw RPCs based on the OST they belong to. Since
grouping RPCs based on their object IDs also requires to take
into consideration the OST They belong to (which ORR does do),
TRR can be thought of as a logical subset of ORR.

A policy can be enabled/disabled separately for regular and
high-priority requests of each PTLRPC service. This is accomplished
via interaction with the nrs_policies lprocfs file for each service.
Strings written to these files need to be of the form "<policy_name>
[start|stop] optional:[reg|hp]", for example:
lctl set_param mdt.*.mdt_setattr.nrs_policies="crr-e start", or
lctl set_param ost.OSS.ost_io.nrs_policies="trr stop reg"

When the optional [reg|hp] token is omitted, the requested change will
take effect on both regular and high-priority RPCs for the service;
please note the use of quoting around the string tokens.

Credit should go to Liang Zhen for putting together much of the core
framework, and Eric Barton for the binary heap implementation that
this patch adds to libcfs.

Signed-off-by: Nikitas Angelinas <nikitas_angelinas@xyratex.com>
Xyratex-bug-id: MRP-73
Change-Id: Ief97e4bad00140f431eadde59de4ad5ab564a963