LU-445 Make previous change backwards compatible via lst stat flag
With the previous patch as is, a 2.3 system running against a 2.2 or 2.1
system will have an invalid timestamp for doing bandwidth calculations.

This patch adds a new flag, "--remoteTs", to "lst stat" which will trigger
the use of the remote timestamps.  If the flag is not given, then the
previous behaviour, using the local timestamp, will be done.

This patch will be set up to change the default from using local timestamps
to using remote timestamp when the Lustre version hits 2.8.

Signed-off-by: Doug Oucharek <doug@whamcloud.com>
Change-Id: If655144de72d3e46ab32acfcfc35ed79f58d00b2