LU-56 lnet: fix a race for in-creating peer
Creating peer will release lock and allocate memory, it's possible
that LNet is shuttingdown and destroyed everything in this small
window (unlikely), so we need to add an extra refcount on peer table
before allocating new peers to keep shutting down thread wait.

This patch also fixed a typo of locking funciton.

Signed-off-by: Liang Zhen <liang@whamcloud.com>
Change-Id: If8945117885e59a6d2de5ad0dff24b41e7a16bba